Remove ExpandedRowHeigh
